Add new AI Horde v2 Swagger Documentation
Add swagger.json file with the AI Horde v2 API documentation which
contains all the details of the endpoints and models. This will allow
for easier integration with the AI Horde as developers can reference
this file to understand how to interact with the API.

The documentation is quite extensive, covering all aspects of the
API including request bodies, responses, authentication, rate limits
and more.